qa_api/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
psycopg2-binary==2.9.9
redis==5.0.8
httpx==0.27.2
python-dotenv==1.0.1
pydantic==2.9.2
kafka-python==2.0.2 